Create hook for extracting partial CreateSLOForm from url state
