+++ x-pack/plugins/observability/public/pages/slo_edit/helpers/process_slo_form_values.ts
@@ -88,7 +88,7 @@ export function transformValuesToUpdateSLOInput(values: CreateSLOForm): UpdateSL
   };
 }
 
-export function transformCreateSLOInputToCreateSLOForm(
+export function transformPartialCreateSLOInputToPartialCreateSLOForm(
   values: Partial<CreateSLOInput>
 ): Partial<CreateSLOForm> {
   return {
